This adds a test that the "From File" is a regular file, so that if you inadvertently pass in the path to a directory, or some other non-regular file, it will throw a runtime error.
git-svn-id: svn+ssh://s1/data/svnroot/wml/futil@4443 f8f6422b-730d-0410-8a4c-f5298e861a9a

+++ futil/FoundryUtilities.cpp
@@ -1023,6 +1023,14 @@ wml::FoundryUtilities::copyFile (string& from, ostream& to)
 {
 	ifstream in;
 
+	// Test that "from" is a regular file
+	if (!FoundryUtilities::regfileExists (from)) {
+		stringstream ee;
+		ee << "FoundryUtilities::copyFile(): FROM file '"
+		   << from << "' is not a regular file";
+		throw runtime_error (ee.str());
+	}
+
 	in.open (from.c_str(), ios::in);
 	if (!in.is_open()) {
 		throw runtime_error ("FoundryUtilities::copyFile(): Couldn't open FROM file");
